Update README.md
As a data analyst with a growing expertise in SQL, Python, and Power BI, I am dedicated to unlocking the potential of data through insightful analysis and visualization. My journey in data science began with a deep interest in databases and SQL, which has now evolved into a broader skill set encompassing data manipulation, analysis, and reporting.